+++ src/main/java/com/contentstack/sdk/Constants.java
@@ -14,7 +14,7 @@ import java.util.logging.Logger;
 public class Constants {
 
     private static final Logger logger = Logger.getLogger(Constants.class.getSimpleName());
-    protected static final String SDK_VERSION = "1.8.0";
+    protected static final String SDK_VERSION = "1.8.1";
     protected static final String ENVIRONMENT = "environment";
     protected static final String CONTENT_TYPE_UID = "content_type_uid";
     protected static final String SYNCHRONISATION = "stacks/sync";